+++ R/helper_hashes.R
@@ -31,7 +31,10 @@ resampling_task_hashes = function(task, resampling, learner = NULL) {
 task_hash = function(task, use_ids, test_ids = NULL, ignore_internal_valid_task = FALSE) {
   # order matters: we first check for test_ids and then for the internal_valid_task
   internal_valid_task_hash = if (!is.null(test_ids)) {
-    # this does the same as task$divide(ids = test_ids)$internal_valid_task$hash but avoids the deep clone
+    # this does the same as 
+    # task$internal_valid_task = test_ids
+    # $internal_valid_task$hash
+    # but avoids the deep clone
     task_hash(task, use_ids = test_ids, test_ids = NULL, ignore_internal_valid_task = TRUE)
   } else if (!ignore_internal_valid_task) {
     task$internal_valid_task$hash