+++ R/mlr_reflections.R
@@ -135,7 +135,7 @@ local({
 
 
   ### Measures
-  tmp = c("na_score", "requires_task", "requires_learner", "requires_model", "requires_train_set")
+  tmp = c("na_score", "requires_task", "requires_learner", "requires_model", "requires_train_set", "primary_iters")
   mlr_reflections$measure_properties = list(
     classif = tmp,
     regr = tmp
@@ -148,7 +148,7 @@ local({
   )
 
   ### ResampleResult
-  mlr_reflections$rr_names = c("task", "learner", "resampling", "iteration", "prediction")
+  mlr_reflections$rr_names = c("task", "learner", "resampling", "iteration")
 
   ### Logger
   mlr_reflections$loggers = list()