+++ tests/testthat/test_mlr_reflections.R
@@ -59,13 +59,13 @@ test_that("resampling works", {
   rr = resample(task, learner, rsmp("cv", folds = 3))
   expect_equal(rr$task_type, "test")
 
-  scores = rr$score(msr("classif.ce"))
-  expect_list(scores$prediction, "Prediction")
+  scores = rr$score(msr("classif.ce"), predictions = TRUE)
+  expect_list(scores$prediction_test, "Prediction")
   expect_numeric(scores$classif.ce, any.missing = FALSE)
   expect_number(rr$aggregate(msr("classif.ce")))
 
-  scores = rr$score()
-  expect_list(scores$prediction, "Prediction")
+  scores = rr$score(predictions = TRUE)
+  expect_list(scores$prediction_test, "Prediction")
   expect_numeric(scores$classif.ce, any.missing = FALSE)
   expect_number(rr$aggregate(msr("classif.ce")))
 })
@@ -86,4 +86,21 @@ test_that("benchmark works", {
 test_that("set column roles works", {
   expect_task(task$set_col_roles("age", "test"))
   expect_equal(task$col_roles$test, "age")
+})
+
+test_that("external packages can set column roles", {
+  x = utils::getFromNamespace("mlr_reflections", ns = "mlr3")
+  old_col_roles = x$task_col_roles$classif
+
+  on.exit({
+    x$col_roles$classif = old_col_roles
+  }, add = TRUE)
+
+  x$task_col_roles$classif = c(x$task_col_roles$classif, "extra_role")
+
+  task = tsk("pima")
+
+  with_future(future::multisession, {
+    resample(task, lrn("classif.rpart"), rsmp("cv", folds = 3))
+  })
 })